requirements.txt
anyascii==0.3.1
apt-clone==0.2.1
apt-xapian-index==0.49
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.8
attrs==21.4.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.9.3
bleach==5.0.1
blinker==1.4
botocore==1.17.60
Brlapi==0.7.0
cached-property==1.5.1
cement==2.8.2
certifi==2022.6.15
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
crudini==0.9.3
cryptography==37.0.4
cupshelpers==1.0
dbus-python==1.2.16
debugpy==1.6.3
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
dirspec==13.10
distlib==0.3.6
distro==1.4.0
Django==4.0.4
django-admin==2.0.1
django-allauth==0.50.0
django-ckeditor==6.4.2
django-crispy-forms==1.14.0
django-debug-toolbar==3.2.4
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-filter==21.1
django-js-asset==2.0.0
django-modelcluster==6.0
django-permissionedforms==0.1
django-six==1.0.4
django-taggit==2.1.0
django-treebeard==4.5.1
djangorestframework==3.13.1
dnspython==2.2.1
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
draftjs-exporter==2.1.7
duplicity==0.8.12.0
email-validator==1.1.3
entrypoints==0.3
et-xmlfile==1.1.0
executing==0.10.0
fasteners==0.14.1
fastjsonschema==2.16.1
feedparser==6.0.10
filelock==3.7.1
Flask==2.0.2
Flask-Admin==1.6.0
Flask-Caching==1.10.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.0
Flask-WTF==1.0.0
future==0.16.0
gpg===1.13.1-unknown
greenlet==1.1.2
grpcio==1.16.1
html5lib==1.1
httplib2==0.14.0
idna==3.3
ifaddr==0.1.6
IMDbPY==6.8
importlib-metadata==4.12.0
iniconfig==1.1.1
iniparse==0.4
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
jmespath==0.10.0
jsonschema==3.2.0
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-pygments==0.2.2
keyring==18.0.1
l18n==2021.3
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mistune==2.0.4
monotonic==1.5
more-itertools==4.2.0
nbclient==0.6.7
nbconvert==7.0.0
nbformat==5.4.0
nemo-emblems==5.2.0
nest-asyncio==1.5.5
netaddr==0.7.19
netifaces==0.10.4
notebook==6.4.12
numpy==1.18.5
oauthlib==3.2.0
olefile==0.46
oneconf==0.3.9
openpyxl==3.0.10
packaging==21.3
PAM==0.4.2
pandocfilters==1.5.0
paramiko==2.6.0
parso==0.8.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==9.2.0
piston-mini-client==0.7.5
platformdirs==2.5.2
pluggy==1.0.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.6.1
psutil==5.5.1
psycopg2==2.8
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.2
pycairo==1.16.2
pycparser==2.21
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.2
Pygments==2.13.0
PyGObject==3.36.0
PyICU==2.4.2
pyinotify==0.9.6
PyJWT==2.4.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.9
pyparted==3.11.2
PyQt5==5.14.1
pyRFC3339==1.1
pyrsistent==0.15.5
PySocks==1.6.8
pytest==7.0.1
python-apt==2.0.1
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-magic==0.4.16
python-xapp==2.2.1
python-xlib==0.23
python3-openid==3.2.0
pytz==2022.1
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==23.2.1
reportlab==3.5.34
requests==2.28.1
requests-file==1.4.3
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
screen==1.0.1
SecretStorage==2.3.1
Send2Trash==1.8.0
sessioninstaller==0.0.0
setproctitle==1.1.10
sgmllib3k==1.0.0
simplejson==3.16.0
sip==4.19.21
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.39
sqlparse==0.4.2
stack-data==0.4.0
system-service==0.3
systemd-python==234
tablib==3.2.1
telepath==0.2
terminado==0.15.0
texttable==1.6.2
thin-client-config-agent==0.8
tinycss2==1.0.2
tldextract==2.2.1
tomli==2.0.1
torbrowser-launcher==0.3.2
tornado==6.2
traitlets==5.3.0
ubuntu-advantage-tools==27.9
ubuntu-drivers-common==0.0.0
ufw==0.36
Unidecode==1.3.4
urllib3==1.26.10
vboxapi==1.0
virtkey==0.63.0
virtualenv==20.19.0
wadllib==1.3.3
wagtail==3.0.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.53.0
Werkzeug==2.0.0
whitenoise==6.0.0
Willow==1.4.1
WTForms==3.0.1
xkit==0.0.0
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
youtube-dl==2021.4.26
zipp==1.0.0
